refactor(control-ui): extract LazyChangeInput into its own module
Move the inline LazyChangeInput component out of the 2000-line index.tsx
into its own file so it can be rendered and tested in isolation, without
pulling in the whole control UI and its font-face side effects. No
behavior change.
